:recycle: refactor : move Bookshelf.js to src
