+++ api/src/certification/shared/infrastructure/repositories/certification-report-repository.js
@@ -1,7 +1,7 @@
 import bluebird from 'bluebird';
 import _ from 'lodash';
 
-import { Bookshelf } from '../../../../../lib/infrastructure/bookshelf.js';
+import { Bookshelf } from '../../../../shared/infrastructure/bookshelf.js';
 import { BookshelfAssessment } from '../../../../shared/infrastructure/orm-models/Assessment.js';
 import { BookshelfCertificationCourse } from '../../../../shared/infrastructure/orm-models/CertificationCourse.js';
 import * as bookshelfToDomainConverter from '../../../../shared/infrastructure/utils/bookshelf-to-domain-converter.js';
