+++ api/src/shared/infrastructure/orm-models/Badge.js
@@ -1,6 +1,6 @@
 import './TargetProfile.js';
 
-import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
+import { Bookshelf } from '../bookshelf.js';
 
 const modelName = 'Badge';
