+++ api/src/shared/infrastructure/orm-models/KnowledgeElement.js
@@ -1,7 +1,7 @@
 import './Assessment.js';
 import './User.js';
 
-import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
+import { Bookshelf } from '../bookshelf.js';
 
 const modelName = 'KnowledgeElement';
