+++ api/src/shared/infrastructure/orm-models/Membership.js
@@ -1,7 +1,7 @@
 import './Organization.js';
 import './User.js';
 
-import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
+import { Bookshelf } from '../bookshelf.js';
 
 const modelName = 'Membership';
