+++ api/src/shared/infrastructure/orm-models/Session.js
@@ -1,6 +1,6 @@
 import './CertificationCandidate.js';
 
-import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
+import { Bookshelf } from '../bookshelf.js';
 
 const modelName = 'Session';
