+++ api/tests/integration/infrastructure/repositories/complementary-certification-course-result-repository_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { knex } from '../../../../lib/infrastructure/bookshelf.js';
 import * as complementaryCertificationCourseResultRepository from '../../../../lib/infrastructure/repositories/complementary-certification-course-result-repository.js';
 import { ComplementaryCertificationCourseResult } from '../../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
+import { knex } from '../../../../src/shared/infrastructure/bookshelf.js';
 import { databaseBuilder, domainBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | Repository | complementary-certification-courses-result-repository', function () {
