+++ api/tests/integration/infrastructure/repositories/complementary-certification-habilitation-repository_test.js
@@ -1,5 +1,5 @@
-import { knex } from '../../../../lib/infrastructure/bookshelf.js';
 import * as complementaryCertificationHabilitationRepository from '../../../../lib/infrastructure/repositories/complementary-certification-habilitation-repository.js';
+import { knex } from '../../../../src/shared/infrastructure/bookshelf.js';
 import { databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | Infrastructure | Repository | complementary-certification-habilitation-repository', function () {
